# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for querying GPU telemetry data",
        "title": "GPU Telemetry Pipeline API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/gpus": {
            "get": {
                "description": "Returns a list of all GPUs for which telemetry data exists. The response includes GPU UUID, device ID, model, and hostname.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GPUs"
                ],
                "summary": "List all GPUs with telemetry",
                "responses": {
                    "200": {
                        "description": "List of GPUs with telemetry data",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.ListGPUsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gpus/{id}/telemetry": {
            "get": {
                "description": "Returns all telemetry entries for a specific GPU, ordered by ingestion time (oldest first). Supports optional time window filters via query parameters. Both start_time and end_time are inclusive.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telemetry"
                ],
                "summary": "Get telemetry for a specific GPU",
                "parameters": [
                    {
                        "type": "string",
                        "example": "GPU-5fd4f087-86f3-7a43-b711-4771313afc50",
                        "description": "GPU UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "2025-01-01T00:00:00Z",
                        "description": "Start time (inclusive, RFC3339 format)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2025-01-01T23:59:59Z",
                        "description": "End time (inclusive, RFC3339 format)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Telemetry records for the GPU",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.GetTelemetryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid GPU UUID or time range",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "GPU not found",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "pkg_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "GPU_NOT_FOUND"
                },
                "error": {
                    "type": "string",
                    "example": "GPU not found"
                },
                "message": {
                    "type": "string",
                    "example": "The requested GPU UUID does not exist"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
        "pkg_api.GPUResponse": {
            "type": "object",
            "properties": {
                "device": {
                    "type": "string",
                    "example": "nvidia0"
                },
                "gpu_id": {
                    "type": "string",
                    "example": "0"
                },
                "hostname": {
                    "type": "string",
                    "example": "mtv5-dgx1-hgpu-031"
                },
                "model": {
                    "type": "string",
                    "example": "NVIDIA H100 80GB HBM3"
                },
                "uuid": {
                    "type": "string",
                    "example": "GPU-5fd4f087-86f3-7a43-b711-4771313afc50"
                }
            }
        },
        "pkg_api.GetTelemetryResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "gpu_uuid": {
                    "type": "string",
                    "example": "GPU-5fd4f087-86f3-7a43-b711-4771313afc50"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_api.TelemetryResponse"
                    }
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "pkg_api.ListGPUsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "gpus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_api.GPUResponse"
                    }
                }
            }
        },
        "pkg_api.TelemetryResponse": {
            "type": "object",
            "properties": {
                "container": {
                    "type": "string"
                },
                "gpu_uuid": {
                    "type": "string",
                    "example": "GPU-5fd4f087-86f3-7a43-b711-4771313afc50"
                },
                "hostname": {
                    "type": "string",
                    "example": "mtv5-dgx1-hgpu-031"
                },
                "ingestion_time": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "metric_name": {
                    "type": "string",
                    "example": "DCGM_FI_DEV_GPU_UTIL"
                },
                "model_name": {
                    "type": "string",
                    "example": "NVIDIA H100 80GB HBM3"
                },
                "namespace": {
                    "type": "string"
                },
                "pod": {
                    "type": "string"
                },
                "value": {
                    "type": "string",
                    "example": "100"
                }
            }
        }
    }
}